+++ releaseNotes.md
@@ -1,3 +1,12 @@
+_August 5th, 2019_  **v2.1.7**
+- APIBlox.NetCore.EventStore.CosmosDb, changed configuration options to allow using a name for the collection id.
+
+Thanks,  
+_Slacquer_  
+
+<br>
+<br>
+
 _August 3rd, 2019_  **v2.1.6**
 - APIBlox.AspNetCore.CommandsQueriesControllersOhMy, missing namespace in template.  Whoops, bad merge!